chore: orbit-store: split backend/contracts.rs into params.rs + traits.… [ORB-10407]
orbit-store: split backend/contracts.rs into params.rs + traits.rs
Agent-Model: claude-sonnet-5
